packages/py-core/pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "re-po"
version = "0.1.0"
description = "Reusable GitHub stats card renderer"
readme = "README.md"
authors = [{name = "akuwuh"}]
requires-python = ">=3.10"
dependencies = [
    "httpx>=0.27.0",
    "pydantic>=2.5",
    "typer>=0.9",
    "jinja2>=3.1",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4",
]

[project.scripts]
"re-po" = "re_po.cli:app"

[tool.setuptools.packages.find]
where = ["src"]

[tool.setuptools.package-data]
"re_po" = ["default_config.yml"]
